+++ meze/functions.py
@@ -284,31 +284,32 @@ def write_slurm_script(template_file, path, log_dir, protocol_file, extra_option
         slurm script 
     """
     output = path + "/" + template_file
-    meze = os.environ["MEZEHOME"]
-    template = meze + "/" + template_file
-    with open(template, "r") as file:
-        lines = file.readlines()
-    
-    if extra_options:
-        options = extra_options
-    elif not extra_options:
-        options = {"PATH_TO_LOGS": log_dir,
-                   "N_TASKS": str(1),
-                   "N_GPUS": str(1), 
-                   "N_CPUS": str(10),
-                #    "MEMORY": str(4069), # only for water and equilibration
-                   "PATH_TO_MEZE": meze,
-                   "PROTOCOLFILE": protocol_file}    
-    if extra_lines:
-        for key, value in extra_lines.items():
-            options[key] = str(value)
+    if not os.path.isfile(output):
+        meze = os.environ["MEZEHOME"]
+        template = meze + "/" + template_file
+        with open(template, "r") as file:
+            lines = file.readlines()
+        
+        if extra_options:
+            options = extra_options
+        elif not extra_options:
+            options = {"PATH_TO_LOGS": log_dir,
+                    "N_TASKS": str(1),
+                    "N_GPUS": str(1), 
+                    "N_CPUS": str(10),
+                    #    "MEMORY": str(4069), # only for water and equilibration
+                    "PATH_TO_MEZE": meze,
+                    "PROTOCOLFILE": protocol_file}    
+        if extra_lines:
+            for key, value in extra_lines.items():
+                options[key] = str(value)
 
-    with open(output, "w") as file:
-        for line in lines:
-            for key, value in options.items():
-                line = line.replace(key, value)
-            file.write(line)
-    os.system(f"chmod +x {output}")
+        with open(output, "w") as file:
+            for line in lines:
+                for key, value in options.items():
+                    line = line.replace(key, value)
+                file.write(line)
+        os.system(f"chmod +x {output}")
     return output
